Automatically download the dictionary file if it does not exist
If we do not find a local copy of the dictionary then we download it
automatically and go from there, because it would be annoying spit out
an error telling the user about using the ‘--download-dictionary’
argument when we can do that for them, i.e. acting as if the program
received that argument.