Add the --match argument to control where the program tries matching
This patch makes it possible to search for Esperanto in three
different places:

1. The beginning of a word.
2. The end of a word.
3. Anywhere in a word

The default behavior is to search only for matches at the beginning of
the words.  Because Esperanto uses suffixes so frequently it seems
like it would be better to match the start of words by default,
allowing the user to more easily find the root of a word.

This patch also removes ‘import sys’ because we no longer need it.