Prepare version 1.0.0 by fixing word matching
This patch correctly the regular expressions built via the ‘--match’
argument as they had some issues, such as a misuse of ‘$’ and a
failure to match letters at the start of words by not allowing
arbitrary characters to follow them, and so on.

The patch also removes ‘--download-dictionary’.  Initially the option
was useful to me.  But when preparing to release the program I
realized it was only useful for testing.  The more I thought about it
the more I came to the opinion that users would never need it.