Show output for exact matches when using the default ‘--match start’
The regular expression used for matching the start of entries requires
letters to follow the search term.  That creates a bug when searching
the dictionary for exact matches because there will be no ‘left-over’
characters to satisfy the regular expression.  For example, searching
for ‘diafana’ prior to this patch produces no output.  With this patch
the program displays the single entry correctly.