Do not download the dictionary when using a symlink for the program
VortSercxilo always looks for the ESPDIC in the same directory where
the user executes the program.  If the user executes the program via a
symlink then it will download the dictionary there even if the
dictionary already exists.  This patch changes the definition of the
global DICTIONARY_FILENAME string so that it looks for the dictionary
in the real location of VortSercxilo, not the directory where the user
runs the program via a symlink.