Refactor and separate the logic for finding dictionary entries
This patch introduces a function for searching the Esperanto
dictionary and collecting the results in a list.  Then it changes the
main loop to use this function and call print() on the results.
Future versions of the program will benefit from having this logic in
a separate function, i.e. we will be able to do more to a list of
matching dictionary entries than print them to screen.