Store all valid match types in a global tuple
There are multiple places in the program where we refer to the ‘valid
match types’ we accept, both in code and in comments.  This patch
moves the list of valid match types into a global tuple so that we
have a more obvious place of reference when mentioning match types.