Implement and use the InvalidMatchType exception
Python’s standard argument parsing library makes sure ‘--match’
receives a valid value on the command-line.  However, we perform no
sanity checks for the value of the ‘match’ parameter of the
collect_matches() function.  This patch creates an exception to use
when we expect to receive the name of a match type and instead receive
something that is not part of VALID_MATCH_TYPES.  The patch also
modifies the collect_matches() function to raise the exception if its
‘match’ parameter has a value we do not support or expect.