Remove all suffixes from words when using ‘--roots-only’
This patch compiles a list of regular expressions to match all
possible suffixes.  If given the ‘--roots-only’ option the program
will first try to remove every suffix from each word before searching
the dictionary.

This process is not bullet-proof.  For example, it will not remove
suffixes if the user only enters part of a word and that part happens
to not match anything in the suffix list.