Remove official prefixes when using ‘--roots-only’
This patch adds a list of prefixes to remove and also modifies the way
that we build the global list of affixes by moving that logic into a
function which we only call if given the ‘--roots-only’ option.  It is
not a big optimization but it does prevent us from wasting time
compiling regular expressions we may not need to use.