[Fix] Show the correct error message for missing entries
If the program could not find an entry it would print an error message
telling the user that there was nothing in the dictionary for the
given `word`.  The problem, however, is that `word` is not the correct
variable to use; that meant whenever the user searches for a term not
in the dictionary, an error occurs.  This patch corrects that problem
by using the appropriate variable for the error message.  Now searching
for an unknown term fails "correctly", or as correctly as a dictionary
can fail, anyway.